Fix compilation issues with gcc 12.2.0.

+++ shared/gimxdownloader/Downloader.h
@@ -26,7 +26,7 @@ public:
     typedef int (*ProgressCallback)(void *clientp, Downloader::DownloaderStatus status, double progress, double total);
     DownloaderStatus download(const std::string& url, std::string& output);
     DownloaderStatus download(const std::string& url, const std::string& file, ProgressCallback callback, void * clientp);
-    int progress(double dlnow, double dltotal);
+    int progress(int64_t  dlnow, int64_t  dltotal);
     static std::string generateTempFile(const std::string& extension);
     static std::string getProgress(double progress, double total);
 private:
